refactor: #38-폴더 구조 변경
